requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pandas
pinecone-client
sentence-transformers
tqdm
langchain
transformers
ollama
jupyters